fixing overflow problems

+++ src/util/math/matrix_base.h
@@ -87,8 +87,8 @@ class Matrix_base : public btas::Tensor2<DataType> {
     Matrix_base<DataType>& operator=(Matrix_base<DataType>&& o);
 
     size_t size() const { return ndim()*mdim(); }
-    int ndim() const { return this->extent(0); }
-    int mdim() const { return this->extent(1); }
+    size_t ndim() const { return this->extent(0); }
+    size_t mdim() const { return this->extent(1); }
 
     void fill_upper();
     void fill_upper_conjg();
